feat(core): Dispatcher + ServiceBase + ErrorSender msg_id uint16→uint32 마이그레이션
- MessageDispatcher: Handler 시그니처, register/unregister/dispatch/has_handler 파라미터 u32
- ServiceBase: handle(), route() 및 내부 람다 msg_id u32, registered_msg_ids_ u32 set
- ErrorSender: build_error_frame() original_msg_id u32, 지정 초기화 필드 순서 v2 반영
- 로깅 포맷 0x{:04x} → 0x{:08x}

+++ apex_core/include/apex/core/error_sender.hpp
@@ -19,7 +19,7 @@ public:
     ///                수신측에서 nullptr 체크 필요)
     /// @return 직렬화된 프레임 바이트 벡터
     [[nodiscard]] static std::vector<uint8_t> build_error_frame(
-        uint16_t original_msg_id,
+        uint32_t original_msg_id,
         ErrorCode code,
         std::string_view message = "");
 };
